version 0.0.7, see changelog
