+++ lib/models/class_member_model.dart
