+++ lib/models/class_model.dart
